+++ docs/content/News.rst
@@ -5,6 +5,7 @@ snakePipes 3.5.0
 ----------------
 
 * Allelic-whatshap mode was added to DNAmapping. Requires --phasedVcf.
+* --useSpikeinForNorm sption to split bam files by host and spikein genome, and to generate spikein-normalized bamCoverage tracks, was added to DNAmapping.
 * Allelic-whatshap mode now works with alignment-free mode and fromBAM in the mRNAseq workflow.